+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Note that Sockeye has checks in place to not translate with an old model that wa
 Each version section may have have subsections for: _Added_, _Changed_, _Removed_, _Deprecated_, and _Fixed_.
 
 ## [1.18.59]
+### Added
 - A flag is returned alongside the best metric indicating if training converged.
 ### Changed
 - Training state cleanup will not be performed for training runs that did not converge.